% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwi.R
\name{mwi}
\alias{mwi}
\title{Download data from KGS Kansas Master Ground-Water Well Inventory by choosing county}
\usage{
mwi(county, unmatched = FALSE)
}
\arguments{
\item{county}{County Name}

\item{unmatched}{Option for downloading unmatched data}
}
\value{
Tibble data
}
\description{
Download data from KGS Kansas Master Ground-Water Well Inventory by choosing county
}
\examples{
# Download data for given county
\dontrun{mwi("allen")}

}
