% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwi_plss.R
\name{mwi_plss}
\alias{mwi_plss}
\title{Download data from KGS Kansas Master Ground-Water Well Inventory by choosing township, section, range}
\usage{
mwi_plss(township, range, range_dir, section = "0", unmatched = FALSE)
}
\arguments{
\item{township}{Township}

\item{range}{Range}

\item{range_dir}{Range Direction}

\item{section}{Section}

\item{unmatched}{Option for downloading unmatched data}
}
\value{
Tibble data
}
\description{
Download data from KGS Kansas Master Ground-Water Well Inventory by choosing township, section, range
}
\examples{
# Download data for given township, range, range direction, and section (optional).
\dontrun{mwi_plss(township = "1", range = "1", range_dir = "W", section = "1", unmatched = TRUE)}

}
