% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogw.R
\name{ogw}
\alias{ogw}
\title{Download data from KGS Master List of Oil and Gas Wells in Kansas by choosing township, range, section, etc.}
\usage{
ogw(township = NULL, range = NULL, range_dir = "W", section = NULL,
  lease = NULL, operator = NULL, state = "kansas",
  county = "any county", api = NULL, welltype = "ALL", all = FALSE)
}
\arguments{
\item{township}{Township}

\item{range}{Range}

\item{range_dir}{Range Direction}

\item{section}{Section}

\item{lease}{Lease Name}

\item{operator}{Operator Name}

\item{state}{State}

\item{county}{County}

\item{api}{5 digit API Well Number}

\item{welltype}{Well Type}

\item{all}{Should we download all data for default values}
}
\value{
Tibble data
}
\description{
Download data from KGS Master List of Oil and Gas Wells in Kansas by choosing township, range, section, etc.
}
\examples{
# Download data for given township, range, section, etc.
\dontrun{ogw(township = "1", range = "1", range_dir = "W", section = "1")}

}
