% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwi.R
\name{mwi}
\alias{mwi}
\title{Download county data from KGS Kansas Master Ground-Water Well Inventory}
\usage{
mwi(county)
}
\arguments{
\item{county}{County Name}
}
\value{
Tibble data
}
\description{
Download county data from KGS Kansas Master Ground-Water Well Inventory
}
\examples{
# Download data for given county
\dontrun{mwi("allen")}

}
